fix(task): dto 수정
